requirements.txt
agate==1.9.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.16.0
cohere==5.3.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
daff==1.3.46
dataclasses-json==0.6.4
dbt-adapters==1.1.1
dbt-common==1.0.4
dbt-core==1.8.0
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.1
fastavro==1.9.4
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.51.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
groq==0.5.0
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.1
grpcio-tools==1.63.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.23.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.16
langchain-cohere==0.1.3
langchain-community==0.0.33
langchain-core==0.1.52
langchain-google-genai==1.0.2
langchain-google-vertexai==1.0.4
langchain-groq==0.1.3
langchain-openai==0.1.3
langchain-pinecone==0.1.1
langchain-text-splitters==0.0.1
langcodes==3.4.0
langgraph==0.0.49
langsmith==0.1.48
language_data==1.2.0
leather==0.4.0
llama-index==0.10.30
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.12
llama-index-core==0.10.30
llama-index-embeddings-openai==0.1.8
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.1
llamaindex-py-client==0.1.18
Logbook==1.5.3
lxml==5.2.2
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mashumaro==3.13
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.23.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
peewee==3.17.5
pexpect==4.9.0
pillow==10.3.0
pinecone-client==3.2.2
platformdirs==4.2.0
portalocker==2.8.2
posthog==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.7.1
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
qdrant-client==1.9.1
rank-bm25==0.2.2
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
sqlparse==0.5.0
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
streamlit-chat==0.1.1
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.0
typer==0.9.4
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
yfinance==0.2.38
zipp==3.18.1